+++ CHANGELOG.md
@@ -73,3 +73,5 @@ significant modifications will be credited to OpenTelemetry Authors.
 ([#281](https://github.com/open-telemetry/opentelemetry-demo/pull/281))
 * Added error for a specific product based on a feature flag
 ([#245](https://github.com/open-telemetry/opentelemetry-demo/pull/245))
+* Added Frontend Instrumentation
+([#293](https://github.com/open-telemetry/opentelemetry-demo/pull/293))